+++ packages/sdk-core/src/subgraph/entities/indexSubscription/indexSubscription.ts
@@ -29,8 +29,10 @@ export interface IndexSubscription {
     updatedAtBlockNumber: BlockNumber;
     approved: boolean;
     indexValueUntilUpdatedAt: BigNumber;
+    indexValueCurrent: BigNumber;
     totalAmountReceivedUntilUpdatedAt: BigNumber;
     units: BigNumber;
+    indexTotalUnits: BigNumber;
     index: SubgraphId;
     token: Address;
     subscriber: Address;
@@ -74,6 +76,8 @@ export class IndexSubscriptionQueryHandler extends SubgraphQueryHandler<
             updatedAtTimestamp: Number(x.updatedAtTimestamp),
             updatedAtBlockNumber: Number(x.updatedAtBlockNumber),
             index: x.index.id,
+            indexValueCurrent: x.index.indexValue,
+            indexTotalUnits: x.index.totalUnits,
             token: x.index.token.id,
             publisher: x.index.publisher.id,
         }));